+++ pages/contactus.vue
@@ -1,109 +1,108 @@
 <template>
   <div>
-    <div style="background-color:black">
-      <div class="container-small overlay p-0" style="background-color:#ededed">
-        <img
-          class="d-block w-100 img1"
-          src="~static/images/contact.png"
-        />
-      </div>
-    </div>
-    <div class="bg-black">
-      <div class="container-small bg-white">
-        <div class="container pad-contact">
-          <h5 style="font-size:12px;margin-bottom: 25px">
-            <nuxt-link style="color: #B71C1C" to="/">HOME </nuxt-link> / CONTACT
-            US
-          </h5>
-          <h2 style="font-weight:700">Write Us</h2>
-          <div class="justify-content-center coloured-rule"></div>
-          <p style="padding: 30px 0;font-size:14px;line-height: 28px">
-            Minimal disruption with complete peace of mind is what our
-            protection services will give you. We like answering questions from
-            our clients. If you have any questions or feel you are ready to
-            begin your next contract, feel free to reach out to us.
-          </p>
-          <div class="row">
-            <div class="col-12 col-sm-12 col-md-6">
-              <div class="d-flex">
-                <img src="/images/clock.png" alt />
-                <p class="m-0 pad">9:30am - 7:30 pm</p>
-              </div>
-              <div class="d-flex">
-                <img src="/images/call-answer.png" alt />
-                <p class="m-0 pad">+91 9152685887</p>
-              </div>
-              <div class="d-flex">
-                <img src="/images/location.png" alt />
-                <p class="m-0 pad">
-                  4th Floor, Protech Centre, Ganeshguri, Ganeshguri, Guwahati -
-                  781006, Above Vishal Mega Mart
-                </p>
-              </div>
-              <div class="d-flex">
-                <img src="images/sent-mail.png" alt />
-                <p class="m-0 pad">pksecurity@gmail.com</p>
-              </div>
-              <div>
-                <a href=""
-                  ><img class="logo-margin" src="/images/twitter.png" alt
-                /></a>
-                <a href=""
-                  ><img class="logo-margin" src="/images/linkedin.png" alt
-                /></a>
-                <a href=""
-                  ><img class="logo-margin" src="/images/pinterest.png" alt
-                /></a>
-                <a href=""
-                  ><img class="logo-margin" src="/images/facebook.png" alt
-                /></a>
-                <a href=""
-                  ><img class="logo-margin" src="/images/gmail.png" alt
-                /></a>
-              </div>
-            </div>
-
-            <div class="col-12 col-sm-12 col-md-6">
-              <div class="information-block">
-                <input
-                  class="primary-input"
-                  type="text"
-                  placeholder="Your Name"
-                  v-model="name"
-                />
-                <input class="primary-input" v-model="email" type="text" placeholder="Email" />
-                <input
-                  class="primary-input"
-                  type="text"
-                  placeholder="Subject"
-                  v-model="subject"
-                />
-                <textarea
-                  class="secondary-input"
-                  type="text"
-                  placeholder="Your Message"
-                  v-model="message"
-                >
-                </textarea>
-              </div>
-              <div class="mt-4 d-flex justify-content-center">
-                <button class="btn btn-send" @click="createContact">
-                  <span>
-                    SEND
-                    <img src="/images/sent-mail-1.png" alt />
-                  </span>
-                </button>
-              </div>
+    <Breadcrumbs title="Contact Us" subtitle crumbs="Home > Contact Us"></Breadcrumbs>
+    <div class="container-small p-3 py-4" style="background-color: #353535;">
+      <p class="contact m-0">Contact Us</p>
+      <span>We're here to help. Let us know if you have any questions at all.</span>
+      <div class="border__btm my-5"></div>
+      <div class="row">
+        <div class="col-12 col-md-6">
+          <span class>LOCATION</span>
+          <div class="d-flex">
+            <i class="mt-3 fa fa-map-marker" aria-hidden="true"></i>
+            <div class="ml-3 mt-3">
+              <span class="d-inline-block">
+                Head Office:
+                4th Floor, Protech Centre, Ganeshguri, Ganeshguri, Guwahati -
+                781006,
+              </span>
+              <span>Administrative Office: House No. 51; 1st Floor, Sanjog Path, Hengrabari Road, Guwahati -781006</span>
             </div>
           </div>
         </div>
+        <div class="col-12 col-md-6" style="line-height:3rem;">
+          <div class="d-flex">
+            <i class="mt-3 fa fa-clock-o" aria-hidden="true"></i>
+            <div class="ml-3">
+              <span>9:30 am - 7:30 pm</span>
+            </div>
+          </div>
+          <div class="d-flex">
+            <i class="mt-3 fa fa-phone" aria-hidden="true"></i>
+            <div class="ml-3">
+              <span>+91 8486027325, +91 8638511195</span>
+            </div>
+          </div>
+          <div class="d-flex">
+            <i class="mt-3 fa fa-location-arrow" aria-hidden="true"></i>
+            <div class="ml-3">
+              <span>pawan@pkservices.in / khatuwalapawan75@gmail.com</span>
+            </div>
+          </div>
+        </div>
+      </div>
+
+      <div class="border__btm my-5"></div>
+      <div class="d-flex justify-content-center mb-5">
+        <iframe
+          src="https://www.google.com/maps/embed?pb=!1m14!1m8!1m3!1d57303.1593547781!2d91.785722!3d26.149553000000004!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x0%3A0x3108e56a1726dc7d!2sPk%20Security%20Service%2Cguwahati!5e0!3m2!1sen!2sus!4v1585831914738!5m2!1sen!2sus"
+          frameborder="0"
+          style="border:0;width: 100%;height: 500px"
+          allowfullscreen
+        ></iframe>
+        
+      </div>
+      <span class="inquiries">General Inquiries</span>
+      <div class="row my-4" style="margin:1.5rem 0!important">
+        <div class="col-12 col-md-4">
+          <span>Full Name:</span>
+        </div>
+        <div class="col-12 col-md-8" style="margin:1.5rem 0!important">
+          <input class="form-control" v-model="name" value />
+        </div>
+      </div>
+      <div class="row my-4" style="margin:1.5rem 0!important">
+        <div class="col-12 col-md-4">
+          <span>Email:</span>
+        </div>
+        <div class="col-12 col-md-8">
+          <input class="form-control" v-model="email" value />
+        </div>
+      </div>
+      <div class="row my-4" style="margin:1.5rem 0!important">
+        <div class="col-12 col-md-4">
+          <span>Phone:</span>
+        </div>
+        <div class="col-12 col-md-8">
+          <input class="form-control" v-model="phone" value />
+        </div>
+      </div>
+      <div class="my-5">
+        <textarea
+          name
+          id
+          cols="146"
+          rows="5"
+          v-model="message"
+          placeholder="How can we help you?"
+          style="width:100%;padding: 10px"
+        ></textarea>
+      </div>
+      <div class="d-flex justify-content-end">
+        <button @click="createContact" class="send__btn" type="submit">SEND</button>
       </div>
     </div>
   </div>
 </template>
 
+
 <script>
+import Breadcrumbs from "@/components/breadcrumbs.vue";
+
 export default {
+  components: {
+    Breadcrumbs
+  },
   data() {
     return {
       name: "",
@@ -140,122 +139,40 @@ export default {
 };
 </script>
 
+
 <style scoped>
-.pad-contact {
-  padding: 40px 0;
+.contact {
+  color: #9e7b5f;
+  font-size: 30px;
+  font-weight: 700;
+  font-style: italic;
 }
-
-.overlay{
-  position: relative;
+.fa {
+  color: #9e7b5f;
+  font-size: 1.1rem;
 }
-
-.overlay:before{
-  position: absolute;
-  content: "";
-  background-color: red;
+span,
+textarea {
+  color: #989595;
 }
-
-img {
-  object-fit: contain;
+.border__btm {
+  border-bottom: 1px solid #989595;
 }
-.pad {
-  padding: 0.63rem;
+.inquiries {
+  color: #9e7b5f;
+  font-size: 1.4rem;
+  font-weight: 600;
 }
-.logo-margin {
-  margin: 0.3rem;
+.form-control,
+textarea {
+  background: #353535;
+  border: 1px solid #989595;
 }
-.primary-input {
-  background: #f6f6f6;
+.send__btn {
+  background-color: #9e7b5f;
   border: none;
-  width: 100%;
-  padding: 0 20px;
-  height: 3rem;
-  margin-bottom: 1.5rem;
-  outline: none;
-}
-.secondary-input {
-  background: #f6f6f6;
-  border: none;
-  padding: 10px 20px;
-  width: 100%;
-  height: 7rem;
-  outline: none;
-}
-.information-block {
-  text-align: end;
-}
-
-.coloured-rule {
-  border-top: 2px solid #B71C1C;
-  width: 5%;
-}
-.align {
-  text-align: center;
-}
-.home {
-  text-decoration: none;
-}
-.para {
-  background-color: #f3f3f3;
-}
-.fonts {
-  font-size: 1.5rem;
-  font-weight: bold;
-}
-.fonts-primary {
-  color: #d82533;
-  font-size: 0.8rem;
-  font-weight: 500;
-}
-.fonts-secondary {
-  color: #221f1f;
-  font-size: 0.8rem;
-  font-weight: 500;
-}
-.fonts-next {
-  font-size: 1.5rem;
-  font-weight: 500;
-}
-
-.btn-send {
-  height: 60px;
-  width: 153px;
-  padding: 9px 49px;
-  font-size: 14px;
-  font-weight: 400;
   color: #fff;
-  border-radius: 0;
-  background: #B71C1C;
-  border: none;
-  text-align: center;
-}
-.btn-send:hover {
-  box-shadow: 0 3px 6px 0 rgba(0, 0, 0, 0.16);
-}
-
-.img1{
-  height: 500px
-}
-
-p{
-  font-size: 14px
-}
-
-@media screen and (max-width: 600px) {
-  .img1{
-    height: 200px
-  }
-  .information-block{
-    padding-top: 20px
-  }
-}
-
-@media screen and (min-width: 601px) and (max-width: 991px){
-  .pad-contact{
-    padding: 40px 0px
-  }
-    .information-block{
-    padding-top: 20px
-  }
+  border-radius: 4px;
+  padding: 0.4rem 1.3rem;
 }
 </style>
