+++ pages/whoweare.vue
@@ -1,107 +1,100 @@
 <template>
-  <div class="bg-black">
-    <div class="container-small p-0 bg-white">
-      <div class="container p-0 pt-5 pb-5">
-        <h5 style="font-size:12px;margin-bottom: 25px">
-          <nuxt-link style="color: #B71C1C" to="/">HOME</nuxt-link>/ WHO WE
-          ARE US
-        </h5>
-        <div>
-          <h3 class="text-left">Who we are</h3>
-          <p class="who-we">
-            PK Security Service is the trusted provider of cutting-edge quality
-            security service. We are committed to providing reliable service
-            through custom-designed security strategies. Gain improved access
-            control to your property with a range of total solutions security
-            package. PK Security service has a reputation of being the top
-            player in providing exclusive security services across Guwahati and
-            North-East. We follow stringent benchmarked standards of security
-            services with a holistic approach towards securing your property.
-            With a firm foothold in the industry, our clientele represents a
-            wide range of specialized industry sectors. Our security services
-            are flexible and encompass all requirements as per our clients
-            dictate. PK Security has always managed to impress clients with a
-            degree of professionalism that is second to none. Our office is
-            based in Guwahati at a prominent location easily accessible via
-            various modes of transport.
-          </p>
+  <div style="background-color:black">
+    <Breadcrumbs title="About Us" subtitle="Unmatchable Quality meets Unbeatable Price Quote" crumbs="Home > Security Companies > About Us"></Breadcrumbs>
+    <div class="container-small">
+      <div class="row mt-5">
+        <div class="col-12 col-lg-8">
+          <div>
+            <p class="text-whoweare">
+              PK Security Service is the trusted provider of cutting-edge quality
+              security service. We are committed to providing reliable service
+              through custom-designed security strategies.
+            </p>
+            <p class="text-whoweare">
+              Gain improved access
+              control to your property with a range of total solutions security
+              package. PK Security service has a reputation of being the top
+              player in providing exclusive security services across Guwahati and
+              North-East.
+            </p>
+            <p class="text-whoweare">
+              We follow stringent benchmarked standards of security
+              services with a holistic approach towards securing your property.
+              With a firm foothold in the industry, our clientele represents a
+              wide range of specialized industry sectors.
+            </p>
+            <p class="text-whoweare">
+              Our security services
+              are flexible and encompass all requirements as per our clients
+              dictate. PK Security has always managed to impress clients with a
+              degree of professionalism that is second to none. Our office is
+              based in Guwahati at a prominent location easily accessible via
+              various modes of transport.
+            </p>
+          </div>
         </div>
-        <div class="middle mt-5">
-          <!-- video section -->
-          <video controls style="height:100%;width:100%">
-            <source src="/video/pkvideo1.mp4" type="video/mp4" />
-          </video>
-        </div>
-
-        <div class="mt-5">
-          <p class="who-we">
-            At PK Security Services, you get a security program and not just a
-            guard. We pledge all of our company resources towards the successful
-            operation and protection of you and your property. Regardless of
-            what industry you're in or what you need to protect, we can provide
-            professional security personnel to meet your needs. We work closely
-            with our clients to establish effective, efficient security services
-            and systems. We are competitive within the industry and believe that
-            our clients deserve reliability and quality. PK Security services a
-            very diverse and discerning clientele. Our service market includes,
-            but is not limited to, commercial office, high rise, corporate,
-            industrial, retail, gated residential communities, and
-            municipalities. Our goal is to provide every clientele with top
-            quality security services available. Our security guards are
-            professionally trained to sense any danger. They are trained in
-            'situation awareness' and 'crime prevention.' In dire times of
-            calamity and conflict, they will always be there to guide the people
-            towards safety.
-          </p>
-        </div>
-
-        <div class="mt-5">
-          <div class="row">
-            <div class="col-12 col-sm-12 col-md-12 col-lg-4 pb-3">
-              <img class="w-100 h-100" src="/images/bgsecurity.jpg" alt />
+        <div class="col-12 col-lg-4">
+          <div>
+            <div class="about-next py-2 position-relative">
+              <div class="about position-absolute">About PK Security & Services</div>
+              <ul class="px-5">
+                <li>About PK Security</li>
+                <li>Corporate Social Responsibility</li>
+                <li>Quality Policy</li>
+                <li>Gallery</li>
+                <li>FAQ</li>
+                <li>Privacy Policy</li>
+              </ul>
             </div>
-            <div class="col-12 col-sm-12 col-md-12 col-lg-4">
+          </div>
+        </div>
+      </div>
+      <div>
+        <div class="row" style="margin-top:50px!important">
+          <div class="col-12 col-sm-12 col-md-12 col-lg-4 pb-3">
+            <img class="w-100 h-100" src="/images/bgsecurity.jpg" alt />
+          </div>
+          <div class="col-12 col-sm-12 col-md-12 col-lg-4">
+            <div>
               <div>
-                <div>
-                  <p class="p-header">MISSION</p>
-                  <p class="p-header-sub">
-                    Our mission is to provide innovative and proactive security
-                    solutions to every client, regardless of the industry. We
-                    want to focus on building a smart and intuitive team of
-                    security personnel who can offer the best solutions during a
-                    time of emergency.
-                  </p>
-                </div>
-                <div>
-                  <p class="p-header">OUR MOTTO</p>
-                  <p class="p-header-sub">
-                    To work towards building a secure environment and protect it
-                    from any sabotage with responsiveness, diligence, and
-                    professionalism.
-                  </p>
-                </div>
+                <p class="p-header">MISSION</p>
+                <p class="p-header-sub">
+                  Our mission is to provide innovative and proactive security
+                  solutions to every client, regardless of the industry. We
+                  want to focus on building a smart and intuitive team of
+                  security personnel who can offer the best solutions during a
+                  time of emergency.
+                </p>
+              </div>
+              <div>
+                <p class="p-header">OUR MOTTO</p>
+                <p class="p-header-sub">
+                  To work towards building a secure environment and protect it
+                  from any sabotage with responsiveness, diligence, and
+                  professionalism.
+                </p>
               </div>
             </div>
-            <div class="col-12 col-sm-12 col-md-12 col-lg-4">
+          </div>
+          <div class="col-12 col-sm-12 col-md-12 col-lg-4">
+            <div>
               <div>
-                <div>
-                  <p class="p-header">VISION</p>
-                  <p class="p-header-sub">
-                    To create a benchmark in retaining the highest rate of
-                    clientele in the contract security industry. We also intend
-                    to assist in the development of each employee while
-                    providing a safe, challenging, and cooperative work
-                    environment.
-                  </p>
-                </div>
-                <div>
-                  <p class="p-header">OUR GOALS</p>
-                  <p class="p-header-sub">
-                    Our future goal is to bring intelligent, automated security
-                    solutions that enable cost-effective, efficient operations
-                    of security monitoring and reporting services.
-                  </p>
-                </div>
+                <p class="p-header">VISION</p>
+                <p class="p-header-sub">
+                  To create a benchmark in retaining the highest rate of
+                  clientele in the contract security industry. We also intend
+                  to assist in the development of each employee while
+                  providing a safe, challenging, and cooperative work
+                  environment.
+                </p>
+              </div>
+              <div>
+                <p class="p-header">OUR GOALS</p>
+                <p class="p-header-sub">
+                  Our future goal is to bring intelligent, automated security
+                  solutions that enable cost-effective, efficient operations
+                  of security monitoring and reporting services.
+                </p>
               </div>
             </div>
           </div>
@@ -110,28 +103,67 @@
     </div>
   </div>
 </template>
+
+
+<script>
+import Breadcrumbs from "@/components/breadcrumbs.vue";
+
+export default {
+  components: {
+    Breadcrumbs
+  }
+};
+</script>
+
 <style scoped>
-.pad-who {
-  padding: 70px 0;
+/* .bg-black-custom {
+} */
+
+.text-whoweare {
+  color: #fff;
+  font-size: 16px;
+  opacity: 0.62;
+  text-align: justify;
 }
 
-.who-we {
-  font-size: 14px;
-  font-weight: normal;
-  font-stretch: normal;
-  font-style: normal;
-  line-height: 28px;
-  letter-spacing: normal;
-  text-align: left;
-  color: #454545;
+.about {
+  background-color: #e5b214;
+  box-shadow: rgba(0, 0, 0, 0.16) !important;
+  width: 80%;
+  margin: auto;
+  margin-top: -2.3rem;
+  height: 60px;
+  border-radius: 3px;
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  font-size: 18px;
+  left: 50%;
+  transform: translateX(-50%);
 }
 
-/* .p-header {
+.about-next {
+  margin-top: 5rem;
+  background-color: #333333;
+}
+
+ul li {
+  list-style: disc inside url(##9e7b5f);
+  color: #fff;
+  opacity: 0.58;
+  margin-top: 20px;
+}
+
+ul li:first-child {
+  margin-top: 50px;
+}
+
+.p-header {
   margin-bottom: 15px;
   font-weight: 700;
   font-size: 15px;
   text-align: left;
-  color: #b71c1c;
+  color: #9e7b5f;
 }
 
 .p-header-sub {
@@ -142,6 +174,7 @@
   line-height: 28px;
   letter-spacing: normal;
   text-align: left;
-  color: #454545;
-} */
+  color: #ffffff;
+  opacity: 0.63;
+}
 </style>
